+++ src/domain/feedbackComponent/types.ts
@@ -1,0 +1,5 @@
+import { FormValues } from "../feedbackForm/types";
+
+export type FeedbackComponentValues = FormValues & {
+  locale?: string;
+};