+++ src/index.ts
@@ -1,3 +1,3 @@
-export { default } from "./domain/feedbackComponent/FeedbackComponent";
+export { default as FeedbackComponent } from "./domain/feedbackComponent/FeedbackComponent";
 export { default as defaultMessages } from "./domain/i18n/defaultMessages";
 export { default as hdsTheme } from "./domain/hdsTheme/hdsTheme";